version 1.0, hay que chequear leaks, recibe correctamente mensajes de texto

+++ trunk/src/dispobject/dispobject.cpp
@@ -107,5 +107,6 @@ DispObject::~DispObject()
 {
 	if (this->user != NULL)
 		delete this->user;
-	deleteFile();
+	if(this->file.exists())
+		deleteFile();
 }
